requirements.txt
channels >= 3.0.3
channels-redis >= 3.2.0
django >= 3.1.3
django-environ >= 0.4.5
django-filter >= 2.4.0
django-redis >= 5.0.0
djangorestframework >= 3.11.1
drf-spectacular >= 0.13.2
psycopg2 >= 2.8.6
pyyaml >= 5.3.1
